Address CI and move code around
